// client/src/typing/auth.ts
export interface LoginForm {
  login: string;
  password: string;
}

export interface BusinesForm {
  fullName: string;
  description: string;
  inn: string;
  ogrn: string;
  address: string;
  email: string;
  phoneNumber: string;
  login: string;
  password: string;
}

export interface StudentForm {
  fullName: string;
  studentIdNumber: string;
  educationalOrganizationFullName: string;
  speciality: string;
  grade: string;
  experienceLevel: string;
  email: string;
  phoneNumber: string;
  login: string;
  password: string;
}